+++ example/dump_sql.R
@@ -8,4 +8,4 @@ rules <- modifier(.file = system.file("db/corrections.yml", package="dcmodifydb"
 print(rules)
 
 # show sql code generated from the rules.
-dump_sql(rules, person)
+dump_sql(rules, person, key = "id")