+++ example/is_working_db.R
@@ -1,4 +1,4 @@
-person <- dbplyr::memdb_frame(age = 12, salary = 3000)
+person <- dbplyr::memdb_frame(id="dirk", age = 12, salary = 3000)
 
 library(dcmodify)
 
@@ -7,10 +7,10 @@ correction_rules <- modifier( if (age < 16) salary = 0
                             )
 
 # second rule is not working, because retired is not available
-is_working_db(correction_rules, person, warn = FALSE)
+is_working_db(correction_rules, person, key="id", warn = FALSE)
 
 # show warnings (default)
-is_working_db(correction_rules, person, warn = TRUE)
+is_working_db(correction_rules, person, key="id", warn = TRUE)
 
 # show the sql statements that are not working
-is_working_db(correction_rules, person, warn = FALSE, sql_warn = TRUE)
+is_working_db(correction_rules, person, key = "id", warn = FALSE, sql_warn = TRUE)