+++ example/person.R
@@ -3,9 +3,11 @@
 library(dcmodify)
 rules <- modifier(.file = system.file("db/corrections.yml", package="dcmodifydb"))
 
-con <- DBI::dbConnect(RSQLite::SQLite(), dbname=system.file("db/person.db", package="dcmodifydb"))
+con <- DBI::dbConnect( RSQLite::SQLite()
+                     , dbname=system.file("db/person.db", package="dcmodifydb")
+                     )
 person <- dplyr::tbl(con, "person")
 print(person)
 
-person2 <- modify(person, rules, copy=TRUE)
+person2 <- modify(person, rules, key="id", copy=TRUE)
 print(person2)